Extract txn stuff from ctx file
The ctx file is now free to accomodate other stuff loaded into the ctx, such as caches
