+++ tests/integration/lens.go
@@ -60,7 +60,7 @@ func configureMigration(
 	_, nodes := getNodesWithIDs(action.NodeID, s.nodes)
 	for _, node := range nodes {
 		txn := getTransaction(s, node.Client, action.TransactionID, action.ExpectedError)
-		ctx := db.SetContextTxn(s.ctx, txn)
+		ctx := db.InitContext(s.ctx, txn)
 
 		err := node.SetMigration(ctx, action.LensConfig)
 		expectedErrorRaised := AssertError(s.t, s.testCase.Description, err, action.ExpectedError)
